+++ doc/changelog.rst
@@ -6,6 +6,7 @@ Changelog
 
 * Add option to set how many OpenMP threads will be used in C extensions
 * Add face area threshold option when computing Voronoi neighbours
+* Add neighbour info window - shows information (separation, etc) about an atoms neighbours
 
 **v141003**